requirements.txt
alembic==1.0.6
arrow==0.13.1
Babel==2.6.0
blinker==1.4
Click==7.0
dateparser==0.7.1
decorator==4.3.2
dominate==2.3.5
elasticsearch==6.1.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
moment==0.8.2
pbr==5.1.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2019.1
regex==2019.4.14
six==1.12.0
SQLAlchemy==1.2.16
sqlalchemy-migrate==0.12.0
sqlparse==0.2.4
Tempita==0.5.2
times==0.7
tzlocal==1.5.1
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
